refactor: merge image collection into remark-resolve-images
Single tree traversal now both resolves and collects image URLs,
eliminating the separate remark-collect-images plugin.
