+++ packages/chronicle/src/server/vite-config.ts
@@ -8,7 +8,6 @@ import path from 'node:path';
 import remarkDirective from 'remark-directive';
 import { type InlineConfig } from 'vite';
 import remarkResolveImages from '../lib/remark-resolve-images';
-import remarkCollectImages from '../lib/remark-collect-images';
 import remarkResolveLinks from '../lib/remark-resolve-links';
 import remarkReadingTime from 'remark-reading-time';
 import remarkUnusedDirectives from '../lib/remark-unused-directives';
@@ -96,7 +95,6 @@ export async function createViteConfig(
               remarkUnusedDirectives,
               remarkResolveLinks,
               remarkResolveImages,
-              remarkCollectImages,
               remarkMdxMermaid,
               remarkReadingTime,
             ],
